extract abstract base for {Hardhat,Fork}Blockchain
